I define my reality
Added a typedef. They make life wonderful
Didn't change everythign into the typedef...

I have a strange note on the line 37 of mymalloc.c I feel like this
would allow us to write past the end of the array or into another
variables data space. I feel that we will end up having a problem with
this in the future while testing. MAKE SURE WE TEST THIS SITUATION.
And maybe, who knows, it won't be so bad because everything is a
pointer...